cloud_functions/climateiq_export_to_aws_cf/requirements.txt
black==24.4.2
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
flake8==7.1.0
Flask==3.0.3
functions-framework==3.7.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.65.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pyflakes==3.2.0
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
